+++ src/edgettsgui/__main__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""python -m edgettsgui"""
+
+from edgettsgui.ui.app import main
+
+if __name__ == "__main__":
+    main()